mm: Add debugfs interface for physical address debugging
This commit introduces a new debugfs interface called
"debug_phys_addr" to the kernel. The interface provides
basic read and write operations for debugging physical
memory addresses.

In debugging scenarios, developers encounter physical addresses
that they need to map back to their corresponding users to understand
which part of the system is using that memory. Currently, there is no
straightforward way to perform this translation directly from physical
addresses to users such as kernel symbols, vmalloc memory, kmalloc
memory, and any processes using that page. This complicates the
debugging process and slows down the resolution of memory-related issues.
This interface is especially useful for debugging and
understanding memory mappings in the kernel. This will allow users to
determine the user of a given physical address.

The debugfs entry is created with the file operations
defined in `debug_phys_addr_proc_ops`. The interface
allows read and write permissions (0600) and is
registered during the kernel's initialization phase
using `device_initcall`.

For this file to be created the CONFIG_DEBUG_PHYS_ADDR needs to be
enabled. Which depends on CONFIG_SLUB_DEBUG and enable slub debugging
as this is required for tracking kmalloc users.

How to use this interface:

 # echo 0x700000 > /sys/kernel/debug/debug_phys_addr
 # cat /sys/kernel/debug/debug_phys_addr
 Mem Space,Physical Addr,Virtual Addr,pid/symbol/vmlinux segment,task name
 kernel symbol,0x700000,0xc000000000700000,pstore_dump+0x410/0x454,kernel

Signed-off-by: Mukesh Kumar Chaurasiya <mchauras@linux.ibm.com>