mm: do rmap_walk to get task information
Writing an hex value to the debug_phys_addr file will save the page
user details to the dynamic buffer.
Leveraging the `rmap_walk` API to iterate over the reverse mappings
of pages, allowing us to identify tasks that have mapped specific pages
in their address space.
This will save the pid and comm of the process into the dynamic buffer
which can be read from the same file.

Signed-off-by: Mukesh Kumar Chaurasiya <mchauras@linux.ibm.com>