mm: get kmalloc requester for a phys addr
To get where the kmalloc memory was allocated, slub already exposes
an interface get_track, which keeps track of the function who requested
the allocation of the kmalloc memory.
Using this we get the address of the caller where the request was made,
the kallsyms_lookup interface translates the addr to the kernel symbol
in the current running kernel.
The size of the temp buffer is increased to 1024 bytes so that it can
accommodate all symbols

Signed-off-by: Mukesh Kumar Chaurasiya <mchauras@linux.ibm.com>